+++ tests/erlang_tests/test_binary_to_term.erl
@@ -27,23 +27,25 @@
     test_atom_decoding/0,
     get_atom/1,
     get_binary/1,
-    test_atom_decoding_checks/0
+    test_atom_decoding_checks/0,
+    id/1
 ]).
 
 start() ->
     % Starting from OTP-26, atoms are encoded as UTF-8 by default.
-    TermToBinaryOptions =
-        case erlang:system_info(machine) of
-            "BEAM" ->
-                case erlang:system_info(version) >= "13.2" of
-                    true -> [{minor_version, 1}];
-                    false -> []
-                end;
-            "ATOM" ->
-                []
-        end,
-    test_reverse(foo, <<131, 100, 0, 3, 102, 111, 111>>, TermToBinaryOptions),
-    test_reverse(bar, <<131, 100, 0, 3, 98, 97, 114>>, TermToBinaryOptions),
+    test_reverse(foo, {<<131, 119, 3, 102, 111, 111>>, <<131, 100, 0, 3, 102, 111, 111>>}),
+    test_reverse(bar, {<<131, 119, 3, 98, 97, 114>>, <<131, 100, 0, 3, 98, 97, 114>>}),
+    test_reverse(
+        '∀x∃y.f(x,y)',
+        <<131, 119, 15, 226, 136, 128, 120, 226, 136, 131, 121, 46, 102, 40, 120, 44, 121, 41>>,
+        []
+    ),
+    test_reverse(
+        ':アトムＶＭ',
+        <<131, 119, 16, 58, 227, 130, 162, 227, 131, 136, 227, 131, 160, 239, 188, 182, 239, 188,
+            173>>,
+        []
+    ),
     test_reverse(128, <<131, 97, 128>>),
     test_reverse(257, <<131, 98, 0, 0, 1, 1>>),
     test_reverse(0, <<131, 97, 0>>),
@@ -51,23 +53,57 @@ start() ->
     test_reverse(32768, <<131, 98, 0, 0, 128, 0>>),
     test_reverse(-32768, <<131, 98, 255, 255, 128, 0>>),
     test_reverse(
-        {foo, bar},
-        <<131, 104, 2, 100, 0, 3, 102, 111, 111, 100, 0, 3, 98, 97, 114>>,
-        TermToBinaryOptions
+        {foo, bar}, {
+            <<131, 104, 2, 119, 3, 102, 111, 111, 119, 3, 98, 97, 114>>,
+            <<131, 104, 2, 100, 0, 3, 102, 111, 111, 100, 0, 3, 98, 97, 114>>
+        }
     ),
-    test_reverse({foo, 0}, <<131, 104, 2, 100, 0, 3, 102, 111, 111, 97, 0>>, TermToBinaryOptions),
+    test_reverse({foo, 0}, {
+        <<131, 104, 2, 119, 3, 102, 111, 111, 97, 0>>,
+        <<131, 104, 2, 100, 0, 3, 102, 111, 111, 97, 0>>
+    }),
     test_reverse([], <<131, 106>>),
     test_reverse(
-        [{foo, 0}, {bar, 1}],
-        <<131, 108, 0, 0, 0, 2, 104, 2, 100, 0, 3, 102, 111, 111, 97, 0, 104, 2, 100, 0, 3, 98, 97,
-            114, 97, 1, 106>>,
-        TermToBinaryOptions
+        [{foo, 0}, {bar, 1}], {
+            <<131, 108, 0, 0, 0, 2, 104, 2, 119, 3, 102, 111, 111, 97, 0, 104, 2, 119, 3, 98, 97,
+                114, 97, 1, 106>>,
+            <<131, 108, 0, 0, 0, 2, 104, 2, 100, 0, 3, 102, 111, 111, 97, 0, 104, 2, 100, 0, 3, 98,
+                97, 114, 97, 1, 106>>
+        }
     ),
     test_reverse(
         [improper | list],
-        <<131, 108, 0, 0, 0, 1, 100, 0, 8, 105, 109, 112, 114, 111, 112, 101, 114, 100, 0, 4, 108,
-            105, 115, 116>>,
-        TermToBinaryOptions
+        {
+            <<131, 108, 0, 0, 0, 1, 119, 8, 105, 109, 112, 114, 111, 112, 101, 114, 119, 4, 108,
+                105, 115, 116>>,
+            <<131, 108, 0, 0, 0, 1, 100, 0, 8, 105, 109, 112, 114, 111, 112, 101, 114, 100, 0, 4,
+                108, 105, 115, 116>>
+        }
+    ),
+    test_reverse({foo, bar}, {
+        <<131, 104, 2, 119, 3, 102, 111, 111, 119, 3, 98, 97, 114>>,
+        <<131, 104, 2, 100, 0, 3, 102, 111, 111, 100, 0, 3, 98, 97, 114>>
+    }),
+    test_reverse({foo, 0}, {
+        <<131, 104, 2, 119, 3, 102, 111, 111, 97, 0>>,
+        <<131, 104, 2, 100, 0, 3, 102, 111, 111, 97, 0>>
+    }),
+    test_reverse([], <<131, 106>>),
+    test_reverse(
+        [{foo, 0}, {bar, 1}], {
+            <<131, 108, 0, 0, 0, 2, 104, 2, 119, 3, 102, 111, 111, 97, 0, 104, 2, 119, 3, 98, 97,
+                114, 97, 1, 106>>,
+            <<131, 108, 0, 0, 0, 2, 104, 2, 100, 0, 3, 102, 111, 111, 97, 0, 104, 2, 100, 0, 3, 98,
+                97, 114, 97, 1, 106>>
+        }
+    ),
+    test_reverse(
+        [improper | list], {
+            <<131, 108, 0, 0, 0, 1, 119, 8, 105, 109, 112, 114, 111, 112, 101, 114, 119, 4, 108,
+                105, 115, 116>>,
+            <<131, 108, 0, 0, 0, 1, 100, 0, 8, 105, 109, 112, 114, 111, 112, 101, 114, 100, 0, 4,
+                108, 105, 115, 116>>
+        }
     ),
     test_reverse(<<"foobar">>, <<131, 109, 0, 0, 0, 6, 102, 111, 111, 98, 97, 114>>),
     test_reverse(<<":アトムＶＭ">>, <<131, 109, 0, 0, 0, 6, 58, 162, 200, 224, 54, 45>>),
@@ -86,7 +122,7 @@ start() ->
             57, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 48,
             49, 50, 51, 52, 53>>
     ),
-    ok = test_external_function(TermToBinaryOptions),
+    ok = test_external_function(),
 
     {32768, 6} = erlang:binary_to_term(<<131, 98, 0, 0, 128, 0, 127>>, [used]),
     test_catenate_and_split([foo, bar, 128, {foo, bar}, [a, b, c, {d}]]),
@@ -99,7 +135,16 @@ start() ->
 test_reverse(T, Interop) ->
     test_reverse(T, Interop, []).
 
-test_reverse(T, Interop, Options) ->
+test_reverse(T, {Utf8Interop, Latin1Interop}, Options) ->
+    case get_otp_version() of
+        X when is_integer(X) andalso X >= 26 ->
+            test_reverse(T, Utf8Interop, Options);
+        atomvm ->
+            test_reverse(T, Utf8Interop, Options);
+        _ ->
+            test_reverse(T, Latin1Interop, Options)
+    end;
+test_reverse(T, Interop, Options) when is_binary(Interop) andalso is_list(Options) ->
     Bin =
         case Options of
             [] -> erlang:term_to_binary(T);
@@ -173,18 +218,27 @@ mutate_bin(Bin, I) ->
     I2 = Ith bxor 16#FF,
     <<Prefix/binary, I2:8/integer-unsigned, Rest/binary>>.
 
-test_external_function(Options) ->
-    T = [fun ?MODULE:apply/2, fun ?MODULE:apply/3],
+test_external_function() ->
+    T = T = [?MODULE:id(fun ?MODULE:apply/2), ?MODULE:id(fun ?MODULE:apply/3)],
     Bin =
-        case Options of
-            [] -> erlang:term_to_binary(T);
-            _ -> erlang:term_to_binary(T, Options)
+        case get_otp_version() of
+            X when is_integer(X) andalso X >= 26 orelse X == atomvm ->
+                %% expect SMALL_ATOM_UTF8_EXT encoding
+                <<131, 108, 0, 0, 0, 2, 113, 119, 19, 116, 101, 115, 116, 95, 98, 105, 110, 97, 114,
+                    121, 95, 116, 111, 95, 116, 101, 114, 109, 119, 5, 97, 112, 112, 108, 121, 97,
+                    2, 113, 119, 19, 116, 101, 115, 116, 95, 98, 105, 110, 97, 114, 121, 95, 116,
+                    111, 95, 116, 101, 114, 109, 119, 5, 97, 112, 112, 108, 121, 97, 3, 106>>;
+            _ ->
+                %% expect ATOM_EXT encoding
+                <<131, 108, 0, 0, 0, 2, 113, 100, 0, 19, 116, 101, 115, 116, 95, 98, 105, 110, 97,
+                    114, 121, 95, 116, 111, 95, 116, 101, 114, 109, 100, 0, 5, 97, 112, 112, 108,
+                    121, 97, 2, 113, 100, 0, 19, 116, 101, 115, 116, 95, 98, 105, 110, 97, 114, 121,
+                    95, 116, 111, 95, 116, 101, 114, 109, 100, 0, 5, 97, 112, 112, 108, 121, 97, 3,
+                    106>>
         end,
-    Bin =
-        <<131, 108, 0, 0, 0, 2, 113, 100, 0, 19, 116, 101, 115, 116, 95, 98, 105, 110, 97, 114, 121,
-            95, 116, 111, 95, 116, 101, 114, 109, 100, 0, 5, 97, 112, 112, 108, 121, 97, 2, 113,
-            100, 0, 19, 116, 101, 115, 116, 95, 98, 105, 110, 97, 114, 121, 95, 116, 111, 95, 116,
-            101, 114, 109, 100, 0, 5, 97, 112, 112, 108, 121, 97, 3, 106>>,
+
+    Bin = erlang:term_to_binary(T),
+
     [Fun2, Fun3] = binary_to_term(Bin),
     true = is_function(Fun2),
     true = is_function(Fun3),
@@ -337,3 +391,14 @@ expect_badarg(Fun) ->
         _:badarg ->
             ok
     end.
+
+get_otp_version() ->
+    case erlang:system_info(machine) of
+        "BEAM" ->
+            list_to_integer(erlang:system_info(otp_release));
+        _ ->
+            atomvm
+    end.
+
+id(X) ->
+    X.